@@ -35,8 +35,10 @@
     "security-check": "./node_modules/.bin/nsp check",
     "test-debug": "node --inspect-brk TZ=UTC node_modules/.bin/jest --watch --config jest.config.js --runInBand --no-cache",
     "unit": "jest --config jest.config.js --ci",
+    "unit:disabled-compat": "DISABLE_JEST_COMPAT_MODE=true jest --config jest.config.js --ci --group=disabledCompat",
     "unit-setup": "npm run generate-component-import-resolver-map",
     "unit-watch": "jest --config jest.config.js --watch",
+    "unit-watch:disabled-compat": "DISABLE_JEST_COMPAT_MODE=true jest --config jest.config.js --watch --group=disabledCompat",
     "vite": "vite",
     "vite:build": "vite build",
     "webpack": "mode=development webpack-dev-server",
@@ -53,9 +55,9 @@
     "@shopware-ag/webpack-copy-after-build": "1.0.1",
     "@shopware-ag/webpack-kill-process-on-failed-build": "1.0.0",
     "@shopware-ag/webpack-plugin-injector": "1.0.6",
-    "@swc/core": "^1.3.101",
-    "@vue/compat": "3.3.4",
-    "@vue/compiler-sfc": "3.3.4",
+    "@swc/core": "1.3.101",
+    "@vue/compat": "3.4.30",
+    "@vue/compiler-sfc": "3.4.30",
     "@vue/component-compiler-utils": "3.3.0",
     "@vue/devtools-api": "6.5.0",
     "@vue/test-utils": "2.3.2",
@@ -92,6 +94,7 @@
     "lodash": "4.17.21",
     "md5-es": "1.8.2",
     "mini-css-extract-plugin": "1.6.2",
+    "mitt": "3.0.1",
     "optional-chaining-codemod": "1.7.0",
     "patch-package": "6.5.1",
     "path-browserify": "^1.0.1",
@@ -112,7 +115,7 @@
     "uuid": "8.3.2",
     "uuidv7": "0.5.3",
     "v-click-outside": "3.2.0",
-    "vue": "3.3.4",
+    "vue": "3.4.30",
     "vue-apexcharts": "1.6.2",
     "vue-i18n": "9.2.2",
     "vue-router": "4.2.2",
@@ -142,7 +145,7 @@
     "@babel/register": "7.21.0",
     "@lhci/cli": "0.11.0",
     "@shopware-ag/eslint-config-base": "2.0.0",
-    "@shopware-ag/jest-preset-sw6-admin": "4.3.2",
+    "@shopware-ag/jest-preset-sw6-admin": "4.3.4",
     "@testing-library/jest-dom": "5.16.5",
     "@types/lodash": "4.14.192",
     "@types/node": "18.15.10",
@@ -188,10 +191,11 @@
     "express": "4.18.2",
     "function-bind": "1.1.1",
     "http-proxy-middleware": "0.21.0",
-    "jest": "29.5.0",
+    "jest": "^29.7.0",
     "jest-environment-jsdom": "29.5.0",
     "jest-jasmine2": "29.5.0",
     "jest-junit": "15.0.0",
+    "jest-runner-groups": "^2.2.0",
     "jest-silent-reporter": "0.5.0",
     "jsdoc": "3.6.11",
     "jsdom": "21.1.1",
@@ -228,7 +232,7 @@
     "vuex": "Vuex is deprecated and will be removed with Shopware 6.8.0"
   },
   "overrides": {
-    "vue": "3.3.4",
+    "vue": "3.4.30",
     "@typescript-eslint/eslint-plugin": "6.12.0"
   }
 }
