@@ -2,6 +2,7 @@
  * @package admin
  */
 
+import { compatUtils } from '@vue/compat';
 import template from './sw-tabs-item.html.twig';
 import './sw-tabs-item.scss';
 
@@ -31,7 +32,7 @@ Component.register('sw-tabs-item', {
 
     inheritAttrs: false,
 
-    inject: ['feature'],
+    inject: ['feature', 'onNewItemActive'],
 
     props: {
         route: {
@@ -126,7 +127,12 @@ Component.register('sw-tabs-item', {
 
     methods: {
         createdComponent() {
-            this.$parent.$on('new-item-active', this.checkIfActive);
+            if (compatUtils.isCompatEnabled('INSTANCE_CHILDREN')) {
+                this.$parent.$on('new-item-active', this.checkIfActive);
+            } else {
+                this.onNewItemActive(this.checkIfActive);
+            }
+
             if (this.active) {
                 this.isActive = true;
             }
