@@ -1,5 +1,7 @@
 import type { PropType } from 'vue';
 import type { TabItem } from '@shopware-ag/meteor-component-library/dist/esm/components/navigation/mt-tabs/mt-tabs';
+// @ts-expect-error
+import { compatUtils } from '@vue/compat';
 import template from './sw-tabs.html.twig';
 
 const { Component } = Shopware;
@@ -41,7 +43,7 @@ Component.register('sw-tabs', {
             return false;
         },
 
-        itemsBackwardCompatible() {
+        itemsBackwardCompatible(): TabItem[] {
             if (this.items) {
                 return this.items;
             }
@@ -113,6 +115,16 @@ Component.register('sw-tabs', {
             // eslint-disable-next-line @typescript-eslint/no-unsafe-return
             return items;
         },
+
+        // eslint-disable-next-line @typescript-eslint/ban-types
+        listeners(): Record<string, Function | Function[]> {
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access
+            if (compatUtils.isCompatEnabled('INSTANCE_LISTENERS')) {
+                return this.$listeners;
+            }
+
+            return {};
+        },
     },
 
     data(): {
@@ -133,12 +145,15 @@ Component.register('sw-tabs', {
 
     methods: {
         getSlots() {
-            const allSlots = {
-                ...this.$slots,
-                ...this.$scopedSlots,
-            };
+            // eslint-disable-next-line @typescript-eslint/no-unsafe-call,@typescript-eslint/no-unsafe-member-access
+            if (compatUtils.isCompatEnabled('INSTANCE_SCOPED_SLOTS')) {
+                return {
+                    ...this.$slots,
+                    ...this.$scopedSlots,
+                };
+            }
 
-            return allSlots;
+            return this.$slots;
         },
 
         mountedComponent() {
