@@ -1,3 +1,4 @@
+import { compatUtils } from '@vue/compat';
 import template from './sw-custom-field-set-renderer.html.twig';
 import './sw-custom-field-set-renderer.scss';
 
@@ -21,6 +22,8 @@ Component.register('sw-custom-field-set-renderer', {
         'feature',
         'repositoryFactory',
     ],
+
+    compatConfig: Shopware.compatConfig,
 
 
     // Grant access to some variables to the child form render components
@@ -357,13 +360,19 @@ Component.register('sw-custom-field-set-renderer', {
                     // replace the fully fetched set
                     this.sets.forEach((originalSet, index) => {
                         if (originalSet.id === newSet.id) {
-                            this.$set(this.sets, index, newSet);
+                            if (compatUtils.isCompatEnabled('INSTANCE_SET')) {
+                                this.$set(this.sets, index, newSet);
+                            } else {
+                                // eslint-disable-next-line vue/no-mutating-props
+                                this.sets[index] = newSet;
+                            }
                         }
                     });
 
                     // remove the set from the currently loading onces and refresh the visible sets
                     this.loadingFields = this.loadingFields.filter(s => s.id !== setId);
-                }).catch(() => {
+                }).catch((error) => {
+                    console.error(error);
                     // in case of error make loading again possible
                     this.loadingFields = this.loadingFields.filter(s => s.id !== setId);
                 });
