@@ -2,6 +2,7 @@ import { MtTabs } from '@shopware-ag/meteor-component-library';
 import type { PropType } from 'vue';
 import type { TabItem } from '@shopware-ag/meteor-component-library/dist/esm/components/navigation/mt-tabs/mt-tabs';
 import template from './mt-tabs.html.twig';
+import type { TabItemEntry } from '../../../state/tabs.store';
 
 
 // Use the compatConfig from the Shopware object and disable all compatibilities
@@ -38,7 +39,7 @@ Shopware.Component.register('mt-tabs', {
     },
 
     computed: {
-        tabExtensions() {
+        tabExtensions(): TabItemEntry[] {
             return Shopware.State.get('tabs').tabItems[this.positionIdentifier] ?? [];
         },
 
