@@ -1,3 +1,4 @@
+import { compatUtils } from '@vue/compat';
 import template from './sw-tree.html.twig';
 import './sw-tree.scss';
 
@@ -47,9 +48,26 @@ Component.register('sw-tree', {
 
     inject: ['feature'],
 
+    compatConfig: Shopware.compatConfig,
+
     provide() {
+        if (compatUtils.isCompatEnabled('INSTANCE_CHILDREN')) {
+            return {
+                getItems: this.getItems,
+            };
+        }
+
         return {
             getItems: this.getItems,
+            startDrag: this.startDrag,
+            endDrag: this.endDrag,
+            moveDrag: this.moveDrag,
+            addSubElement: this.addSubElement,
+            addElement: this.addElement,
+            duplicateElement: this.duplicateElement,
+            onFinishNameingElement: this.onFinishNameingElement,
+            onDeleteElements: this.onDeleteElements,
+            abortCreateElement: this.abortCreateElement,
         };
     },
 
@@ -268,6 +286,7 @@ Component.register('sw-tree', {
                 this.treeItems = this.getTreeItems(this.isSearched ? null : this.rootParentId);
                 this._eventFromEdit = null;
             },
+            deep: true,
         },
 
         activeTreeItemId(val) {
@@ -650,7 +669,11 @@ Component.register('sw-tree', {
                 return;
             }
 
-            if (typeof this.$listeners['batch-delete'] === 'function') {
+            const batchDeleteIsFunction = compatUtils.checkCompatEnabled('INSTANCE_LISTENERS')
+                ? typeof this.$listeners['batch-delete'] === 'function'
+                : typeof this.$attrs.onBatchDelete === 'function';
+
+            if (batchDeleteIsFunction) {
                 this.$emit('batch-delete', this.checkedElements);
             } else {
                 Object.values(this.checkedElements).forEach((itemId) => {
@@ -672,13 +695,21 @@ Component.register('sw-tree', {
                 if (item.childCount > 0) {
                     this.checkedElementsChildCount += 1;
                 }
-                this.$set(this.checkedElements, item.id, item.id);
+                if (compatUtils.checkCompatEnabled('INSTANCE_SET')) {
+                    this.$set(this.checkedElements, item.id, item.id);
+                } else {
+                    this.checkedElements[item.id] = item.id;
+                }
                 this.checkedElementsCount += 1;
             } else {
                 if (item.childCount > 0) {
                     this.checkedElementsChildCount -= 1;
                 }
-                this.$delete(this.checkedElements, item.id);
+                if (compatUtils.checkCompatEnabled('INSTANCE_DELETE')) {
+                    this.$delete(this.checkedElements, item.id);
+                } else {
+                    delete this.checkedElements[item.id];
+                }
                 this.checkedElementsCount -= 1;
             }
 
