@@ -1,8 +1,11 @@
 /**
  * @package admin
+ * @group disabledCompat
  */
 
 import { mount } from '@vue/test-utils';
+import { mtTextField, mtCheckbox } from '@shopware-ag/meteor-component-library';
+// eslint-disable-next-line import/no-named-as-default,import/no-named-as-default-member
 import getTreeItems from './fixtures/treeItems';
 
 async function createWrapper() {
@@ -18,16 +21,25 @@ async function createWrapper() {
                 'sw-base-field': await wrapTestComponent('sw-base-field'),
                 'sw-checkbox-field': await wrapTestComponent('sw-checkbox-field'),
                 'sw-checkbox-field-deprecated': await wrapTestComponent('sw-checkbox-field-deprecated', { sync: true }),
+                'mt-checkbox': mtCheckbox,
                 'sw-text-field': await wrapTestComponent('sw-text-field'),
+                'mt-text-field': mtTextField,
                 'sw-text-field-deprecated': await wrapTestComponent('sw-text-field-deprecated', { sync: true }),
+                'sw-confirm-field': await wrapTestComponent('sw-confirm-field'),
                 'sw-field-error': true,
                 'sw-tree-input-field': true,
                 'sw-button': true,
                 'sw-context-menu-item': true,
                 'sw-context-button': true,
-                'sw-vnode-renderer': await wrapTestComponent('sw-vnode-renderer'),
+                'sw-vnode-renderer': await wrapTestComponent('sw-vnode-renderer', { sync: true }),
                 'sw-icon': true,
                 'sw-tree-item': await wrapTestComponent('sw-tree-item'),
+                'sw-skeleton': await wrapTestComponent('sw-skeleton'),
+                'sw-inheritance-switch': true,
+                'sw-ai-copilot-badge': true,
+                'sw-help-text': true,
+                'sw-field-copyable': true,
+                'sw-skeleton-bar': true,
             },
             mocks: {
                 $route: {
@@ -38,11 +50,6 @@ async function createWrapper() {
             },
             provide: {
                 validationService: {},
-            },
-            directives: {
-                tooltip: {},
-                droppable: {},
-                draggable: {},
             },
         },
     });
@@ -74,6 +81,7 @@ describe('src/app/component/tree/sw-tree', () => {
         await flushPromises();
 
         await wrapper.get('.sw-tree-item .sw-tree-item__toggle').trigger('click');
+        await flushPromises();
 
         // parent should be open
         const openedParent = wrapper.find('.sw-tree-item.is--opened');
@@ -182,6 +190,8 @@ describe('src/app/component/tree/sw-tree', () => {
         await wrapper.setData({
             checkedElementsCount: 2,
         });
+
+        await flushPromises();
 
         expect(wrapper.find('.sw-tree-actions__delete_categories').attributes().disabled).toBeUndefined();
     });
