@@ -1,4 +1,3 @@
-import Vue from 'vue';
 import Store from 'src/app/store';
 
 /**
@@ -6,10 +5,14 @@ import Store from 'src/app/store';
  * @private
  */
 export default function initStore() {
+    const app = Shopware.Application?.view?.app;
+
     /**
      * This code does two things:
      * 1. Initializing the Pinia singleton by accessing the instance getter.
      * 2. Registering the Pinia plugin with Vue before the first store is trying to be registered.
      */
-    Vue.use(Store.instance._rootState);
+    if (app) {
+        app.use(Store.instance._rootState);
+    }
 }
